Distinguish between external and internal oscillators
There is a subtle distinction between bootloaders for MCUs driven
by an *external* oscillator, which tend to have a small (< 0.2%)
deviation from the nominal F<sub>CPU</sub>, and MCUs running on an
*internal* oscillator, which more often than not exhibit a
deviation of 2% or more from the nominal clock frequency. For the
former, the pre-compiled urboot bootloaders tend to be shorter and
faster at the cost of potentially higher quantisation errors for
the baud rate, up to 2.1%. For the latter the bootloaders tend to
be slower but have a baud rate quantisation error of less than
0.5%; there are a number of bootloaders on offer in steps of 1%
around the nominal F<sub>CPU</sub> so that a particular board can
be furnished with a suitable bootloader for the *actual* internal
clock frequency. All in all the aim is to ensure baud rate errors
are below 2.5% to facilitate reliable serial communication.